// atlas-finance-ui/src/types/companyOverview.ts
export interface CompanyOverview {

  symbol: string;

  companyName: string;

  exchange: string;

  sector: string;

  industry: string;

  country: string;

  currency: string;

  description: string;

  ceo: string;

  employees: number;

  ipoDate: string;

  website: string;

  phone: string;

  address: string;

  city: string;

  state: string;

  zip: string;

  marketCap: number;

  enterpriseValue: number;

  sharesOutstanding: number;

  averageVolume: number;

  beta: number;

  dividendYield: number;

  high52Week: number;

  low52Week: number;
}